Random_Forest/requirements.txt
Flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
matplotlib==3.4.3
Werkzeug==2.0.3
gunicorn==20.1.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
python-dotenv==1.0.0
flask-cors==4.0.0